ticket-service/src/main/java/com/epam/training/ticketservice/core/security/adapter/AccountToUserDetailsAdapter.java
package com.epam.training.ticketservice.core.security.adapter;

import com.epam.training.ticketservice.core.account.entity.Account;
import com.epam.training.ticketservice.core.security.service.SecurityService;
import lombok.Data;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;

import java.util.ArrayList;
import java.util.Collection;

@RequiredArgsConstructor
@Data
public class AccountToUserDetailsAdapter implements UserDetails {

    private final Account account;

    @Override
    public Collection<? extends GrantedAuthority> getAuthorities() {
        var authorities = new ArrayList<SimpleGrantedAuthority>();
        authorities.add(new SimpleGrantedAuthority(SecurityService.NON_PRIVILEGED_ROLE_NAME));
        if (account.isPrivileged()) {
            authorities.add(new SimpleGrantedAuthority(SecurityService.PRIVILEGED_ROLE_NAME));
        }
        return authorities;
    }

    @Override
    public String getPassword() {
        return account.getPassword();
    }

    @Override
    public String getUsername() {
        return account.getUsername();
    }

    @Override
    public boolean isAccountNonExpired() {
        return true;
    }

    @Override
    public boolean isAccountNonLocked() {
        return true;
    }

    @Override
    public boolean isCredentialsNonExpired() {
        return true;
    }

    @Override
    public boolean isEnabled() {
        return true;
    }
}
